Moved expression parser helpers to an internal header to avoid confusing cling
