+++ Framework/Core/include/Framework/Expressions.h
@@ -10,9 +10,7 @@
 #ifndef O2_FRAMEWORK_EXPRESSIONS_H_
 #define O2_FRAMEWORK_EXPRESSIONS_H_
 
-#include "Framework/Kernels.h"
-#include "Framework/VariantHelpers.h"
-
+#include <arrow/compute/kernel.h>
 #include <variant>
 #include <string>
 #include <memory>
@@ -55,26 +53,6 @@ struct BinaryOpNode {
   };
   BinaryOpNode(Op op_) : op{op_} {}
   Op op;
-};
-
-struct ArrowDatumSpec {
-  // datum spec either contains an index, a value of a literal or a binding label
-  std::variant<std::monostate, size_t, LiteralNode::var_t, std::string> datum;
-  explicit ArrowDatumSpec(size_t index) : datum{index} {}
-  explicit ArrowDatumSpec(LiteralNode::var_t literal) : datum{literal} {}
-  explicit ArrowDatumSpec(std::string binding) : datum{binding} {}
-  ArrowDatumSpec() : datum{std::monostate{}} {}
-};
-
-bool operator==(ArrowDatumSpec const& lhs, ArrowDatumSpec const& rhs);
-
-std::ostream& operator<<(std::ostream& os, ArrowDatumSpec const& spec);
-
-struct ArrowKernelSpec {
-  std::unique_ptr<arrow::compute::OpKernel> kernel = nullptr;
-  ArrowDatumSpec left;
-  ArrowDatumSpec right;
-  ArrowDatumSpec result;
 };
 
 /// A generic tree node
@@ -151,9 +129,6 @@ struct Filter {
 
   std::unique_ptr<Node> node;
 };
-
-std::vector<ArrowKernelSpec> createKernelsFromFilter(Filter const& filter);
-
 } // namespace o2::framework::expressions
 
 #endif // O2_FRAMEWORK_EXPRESSIONS_H_
